refactor(settings): extract Map into its own self-contained section
Move the Map section (backend picker, Mapbox/Google Maps blocks, the
match-theme toggle, OSM 3D/Terrain/color-scheme rows, camera rows, and
the three credential dialogs) out of SettingsScreen.kt into a new
self-contained components/MapSection.kt, continuing the per-section
file split started in Tasks 4-6.

MapThemeMatchRow (definition + call) and the OSM map-scheme rows move
here too even though they read as "appearance" content: they were
still physically inside the same Map SettingsSection block as
everything else this task extracts, so the whole block moves verbatim
rather than being partially reimplemented. This is a deliberately
documented intermediate state (see the "NOTE FOR TASK 8" comments) —
the next task relocates that content again, into a new
AppearanceSection.kt, once the old Display section is also split out.

Pure structural move: no behavior change, no test changes. The three
dialog-visibility flags and the dialogs themselves are now hoisted
inside MapSection instead of SettingsScreen, but the composed UI tree
and dispatched actions are identical.
